+++ changelog.md
@@ -16,6 +16,12 @@
   - s12N
   - s2
   - s2N
+* Added the following CIM classes/enums:
+  * `TransformerConstructionKind`
+  * `TransformerFunctionKind`
+* Added the following `PowerTransformer` fields:
+  * `construction_kind: TransformerConstructionKind`
+  * `function: TransformerFunctionKind`
 
 ##### Enhancements
 * None.